Server/requirements.txt
python-magic
python-libmagic

# Ubuntu
libmagic1

# Windows
python-magic-bin
python-magic-win64

# Mac
libmagic
file
